adds query params and filter logic for get-request for products

+++ sql/queries/products.sql
@@ -13,8 +13,8 @@ INSERT INTO products(
 SELECT * FROM products
 WHERE 
   ($1::TEXT = '' OR name ILIKE '%' || $1 || '%') AND
-  ($2 IS NULL OR price >= $2) AND
-  ($3 IS NULL OR price <= $3) AND
+  ($2::NUMERIC IS NULL OR price >= $2) AND
+  ($3::NUMERIC IS NULL OR price <= $3) AND
   ($4::TEXT = '' OR category = $4)
 ORDER BY name;
